Activity 1 and 2

+++ Activity1/Operators.cs
@@ -6,26 +6,25 @@ namespace Activity1
 {
     class Operators
     {
-        public void Result(int ans)
+        public void Result(double ans)
         {
             Console.WriteLine("Result: {0} \n", ans);
         }
-        public void Add(int num1, int num2)
+        public void Add(double num1, double num2)
         {
             Result(num1 + num2);
         }
-
-        public void Subtract(int num1, int num2)
+        public void Subtract(double num1, double num2)
         {
             Result(num1 - num2);
         }
 
-        public void Divide(int num1, int num2)
+        public void Divide(double num1, double num2)
         {
             Result(num1 / num2);
         }
 
-        public void Multiply(int num1, int num2)
+        public void Multiply(double num1, double num2)
         {
             Result(num1 * num2);
         }
